Align frontend types and UI with backend schema
Update frontend to match backend model and improve UI/date handling. Key changes:
- services/api.ts: adjust interfaces (Campanha, Alerta, LogAtividade, DashboardStats) to reflect new DB fields (categoria, dataPostada, usuario object, optional fields) and add basic response checks.
- Alertas, Campanhas, CampanhaDetalhes, AlertaUrgente, Index, Usuarios, Configuracoes: adapt components to new properties (use alerta.categoria, dataPostada, optional date handling), convert campaign form dates to ISO strings, improve null/undefined guards, format dates for pt-BR, tweak icons and small UI/layout refinements, and clean up inline dev comments.
- Usuarios: streamline forms, dropdowns and search input formatting, adjust edit password behaviour and default values.
- Configuracoes: avoid overwriting form when backend returns empty data and add console message on error.
- Add .idea/dataSources.xml (IDE datasource) and include updated build artifacts/Gradle outputs generated by compilation.

These changes align the admin UI with the updated API schema and guard against missing/null data while polishing several UX details.

+++ conecta-a-vida-api/src/main/java/br/edu/ifsp/conectaavida/domain/Usuario.java
@@ -8,14 +8,15 @@ import lombok.*;
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor
 public class Usuario {
 
-    // ATENÇÃO EQUIPE: ID como Integer por causa da tabela existente no Supabase.
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id; // ID atualizado para Long (bigint)
 
     @Column(nullable = false) private String nome;
 
-    // unique = true avisa ao banco para não aceitar dois usuários com o mesmo e-mail.
-    @Column(nullable = false, unique = true) private String email;
+    @Column(nullable = false, unique = true)
+    private String email;
+
     @Column(nullable = false) private String senha;
 
     private Integer idade;
